refactor(quiz): streamline quiz navigation and encapsulate logic in ChoicesSection
- Removed handleSolved function from QuizScreen and integrated its logic into ChoicesSection.
- Updated ChoicesSection to manage quiz navigation directly, enhancing modularity.
- Adjusted button press handling to utilize the new navigation logic.
