+++ app/(tabs)/quiz-tab/quiz/[quizId].tsx
@@ -1,27 +1,18 @@
-import { router, useLocalSearchParams } from 'expo-router'
+import { useLocalSearchParams } from 'expo-router'
 import { SafeAreaView, ScrollView, StyleSheet } from 'react-native'
 
 import { Colors } from '@/constants/Colors'
 import { ChoicesSection } from '@/features/answer-quiz/components/ChoicesSection'
 import { QuestionPrompt } from '@/features/answer-quiz/components/QuestionPrompt'
 import { QuizHeader } from '@/features/answer-quiz/components/QuizHeader'
-import { useNextQuiz } from '@/features/answer-quiz/hooks/useNextQuiz'
 import { useQuizQuery } from '@/features/answer-quiz/hooks/useQuizQuery'
 
 export default function QuizScreen() {
-  const { getNextQuiz } = useNextQuiz()
   const { quizId } = useLocalSearchParams()
 
   if (typeof quizId !== 'string') throw new Error('Invalid quizId')
 
   const { data: quiz } = useQuizQuery(quizId)
-
-  const handleSolved = () => {
-    console.log('handleSolved')
-    const next = getNextQuiz()
-    console.log('next', next)
-    router.push(next ? `/quiz-tab/quiz/${next}` : '/quiz-tab/result')
-  }
 
   if (!quiz) return null
 
@@ -30,7 +21,7 @@ export default function QuizScreen() {
       <SafeAreaView style={styles.safeAreaView}>
         <QuizHeader />
         <QuestionPrompt prompt={quiz.prompt} />
-        <ChoicesSection quiz={quiz} onSolved={handleSolved} />
+        <ChoicesSection quiz={quiz} />
       </SafeAreaView>
     </ScrollView>
   )
